% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cholupdate}
\alias{cholupdate}
\title{Cholesky rank-1 update (downdate)}
\usage{
cholupdate(RMat, xVec, nu)
}
\arguments{
\item{RMat}{square matrix, cholesky decomposition (lower) of matrix A}

\item{xVec}{numeric vector such that \code{length(xVec) = dim(RMat)[j]}, \code{j=1,2}}

\item{nu}{real, positive for update, negative for downdate. Mind the square root in description.}
}
\description{
Let RMat be the lower triangular Cholesky factor of PMat. This function calculates the lower-triangular Cholesky factor of P1 = P + nu^(0.5) * xVec \%*\% t(xVec). For downdate take nu < 0.
}
\details{
This follows the definition in van der Merwe PhD Thesis ``Sigma-Point Kalma Filters for Probabilistic Inference in Dynamic State-Space Models''
}

