% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{filtTests}
\alias{filtTests}
\alias{testUKFclass}
\alias{testSqrtUKFclass}
\title{Filtering application example}
\usage{
testUKFclass(Nlength)

testSqrtUKFclass(Nlength)
}
\arguments{
\item{Nlength}{Integer. Length of generated sample for filtering.}
}
\value{
List with true and estimated states, and the state covariance matrices.
}
\description{
These functions are provided as (1) a simple testing ground for the filter, (2) as a demo of how to write the Transition and Observation handler functions.
}
\details{
The states follow an AR(1) model with correlated errors. All paramters are hard-coded in this example but can be passed in despective \code{Rcpp::List} objects. The observation equations are signed powers of the states, observed with additional correlated error. \code{testUKFclass} runs the standard-form filter. \code{testSqrtUKFclass} runs the square-root form (better numerical stability).
}
