% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ukfMeanCov}
\alias{ukfMeanCov}
\alias{unscentedMean}
\alias{unscentedCov}
\alias{unscentedCrossCov}
\title{Unscented Kalman Filter: Means and Covariances}
\usage{
unscentedMean(xSigma, unscWts)

unscentedCov(xSigma, unscWtsMean, unscWtsCov)

unscentedCrossCov(xSigma, ySigma, unscWtsMean, unscWtsCov)
}
\arguments{
\item{xSigma}{Numeric matrix. Output from \code{\link{generateSigmaPoints}}: expanded state values.}

\item{unscWts}{Numeric vector. Weights for mean calculation. First column from output of \code{\link{generateSigmaWeights}}.}

\item{unscWtsMean}{as above}

\item{unscWtsCov}{Numeric vector. Weights for covariance calculation. Second column from output of \code{\link{generateSigmaWeights}}.}

\item{ySigma}{as above}
}
\value{
Mean vector or (variance-)covariance matrix.
}
\description{
Functions for calculating means and covariances of non-linear transformations of random variables with the use of the Scaled Unscented Transformation.
}
