% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ukfPtsWts}
\alias{generateSigmaPoints}
\alias{generateSigmaWeights}
\alias{ukfPtsWts}
\title{UKF: sigma points and weights}
\usage{
generateSigmaPoints(xMat, gam, PMatChol, L)

generateSigmaWeights(L, alpha, beta)
}
\arguments{
\item{xMat}{Original state matrix: each column is a value of state variables. If \code{ncol(xMat)==1}, the matrix is simply expanded with \code{+/- gam * PMatChol[,k]}. If \code{xMat} has more than one column, it's rewritten and an expansion of the first column is concatenated on the right.}

\item{gam}{Numeric. \code{gamma} parameter for Unscented Transformations.}

\item{PMatChol}{Cholesky decomposition (preferably lower) of the variance-covariance matrix of \code{xMat[,1]}. If past states are propagated, then the top-left block of \code{PMatChol} contains the decomposition and other entries are \code{0}.}

\item{L}{Dimension of state expansion. Should be equal to the number of non-zero entries on \code{diag(PMatChol)}}

\item{alpha}{Numeric in \code{[0,1]}: size of step in the Scaled Uncented Transformation.}

\item{beta}{Numeric. Corrections in weights for calculating higher moments. Set to \code{0} unless you know what you're doing.}
}
\value{
Matrix of size \code{nrow(xMat) x (ncol(xMat) + 2*L)}.

Matrix of size \code{(2*L+1) x 2}. First column contains weights for the calculation of the non-linearly transformed mean; second column -- for the variance-covariance matrix.
}
\description{
Functions for generating sigma-points from original state and a Cholesky decomposition of the state variance-covariance matrix, and for generating weights for calculating mean and covariance via the Unscented Transformation.
}
\section{Functions}{
\itemize{
\item \code{generateSigmaWeights}: 
}}

