\name{ukfRcpp-package}
\alias{ukfRcpp-package}
\alias{ukfRcpp}
\docType{package}
\title{
\packageTitle{ukfRcpp}
}
\description{
\packageDescription{ukfRcpp}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{ukfRcpp}
\packageIndices{ukfRcpp}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{ukfRcpp}

Maintainer: \packageMaintainer{ukfRcpp}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
